use api in order to save notepad's color
