+++ src/components/ColorBox.jsx
@@ -1,11 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { useDispatch } from 'react-redux';
+import { getNotepads as reloadNotepads, updateNotepad } from 'redux/middlewares/notepadMiddlewares';
+
 import 'styles/ColorBox.scss';
 
-const ColorBox = ({ color, setColor }) => {
-  const swapColor = (event) => {
+const ColorBox = ({ notepadId, color }) => {
+  const dispatch = useDispatch();
+
+  const swapColor = async (event) => {
     if (event.keyCode === 229 || event.type === 'click') {
-      setColor(color);
+      const response = await dispatch(updateNotepad({ notepadId, color }));
+      if (response) { await dispatch(reloadNotepads()); }
     }
   };
 
@@ -32,5 +38,5 @@ ColorBox.defaultProps = {
 
 ColorBox.propTypes = {
   color: PropTypes.string,
-  setColor: PropTypes.func.isRequired,
+  notepadId: PropTypes.number.isRequired,
 };
