+++ src/components/NotepadCard.jsx
@@ -1,46 +1,29 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+
 import { Link } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 
 import Card from 'react-bootstrap/Card';
 
 import ColorDropdown from 'components/ColorDropdown';
-import NotepadModal from 'components/NotepadModal';
 import PencilIcon from 'components/PencilIcon';
+import NotepadModal from 'components/NotepadModal';
 import TrashIcon from 'components/TrashIcon';
 
+import notepadImages from 'helpers/notepadImages';
+
 import { destroyNotepad, getNotepads as reloadCurrentNotepads } from 'redux/middlewares/notepadMiddlewares';
-import blueNotepadImage from 'assets/blueNotepadImage.png';
-import redNotepadImage from 'assets/redNotepadImage.png';
-import orangeNotepadImage from 'assets/orangeNotepadImage.png';
-import cyanNotepadImage from 'assets/cyanNotepadImage.png';
-import greenNotepadImage from 'assets/greenNotepadImage.png';
-import yellowNotepadImage from 'assets/yellowNotepadImage.png';
 
 const NotepadCard = ({ notepad }) => {
   const dispatch = useDispatch();
   const [show, setShow] = useState(false);
-  const [color, setColor] = useState('blue');
-  const [notepadImage, setNotepadImage] = useState(blueNotepadImage);
-
-  const notepadImages = {
-    red: redNotepadImage,
-    blue: blueNotepadImage,
-    green: greenNotepadImage,
-    cyan: cyanNotepadImage,
-    yellow: yellowNotepadImage,
-    orange: orangeNotepadImage,
-  };
+  const notepadImage = notepadImages[notepad.color ? notepad.color : 'blue'];
 
   const removeNotepad = async () => {
     const response = await dispatch(destroyNotepad({ notepadId: notepad.id }));
     if (response) { dispatch(reloadCurrentNotepads()); }
   };
-
-  useEffect(() => {
-    setNotepadImage(notepadImages[color]);
-  }, [color]);
 
   return (
     <>
@@ -69,7 +52,7 @@ const NotepadCard = ({ notepad }) => {
               removeNotepad={removeNotepad}
             />
 
-            <ColorDropdown setColor={setColor} />
+            <ColorDropdown notepadId={notepad.id} />
           </div>
 
         </Card.Body>
@@ -91,6 +74,7 @@ export default NotepadCard;
 
 NotepadCard.propTypes = {
   notepad: PropTypes.shape({
+    color: PropTypes.string,
     id: PropTypes.number,
     title: PropTypes.string,
   }).isRequired,
