+++ src/helpers/notepadImages.js
@@ -1,0 +1,17 @@
+import blueNotepadImage from 'assets/blueNotepadImage.png';
+import redNotepadImage from 'assets/redNotepadImage.png';
+import orangeNotepadImage from 'assets/orangeNotepadImage.png';
+import cyanNotepadImage from 'assets/cyanNotepadImage.png';
+import greenNotepadImage from 'assets/greenNotepadImage.png';
+import yellowNotepadImage from 'assets/yellowNotepadImage.png';
+
+const notepadImages = {
+  red: redNotepadImage,
+  blue: blueNotepadImage,
+  green: greenNotepadImage,
+  cyan: cyanNotepadImage,
+  yellow: yellowNotepadImage,
+  orange: orangeNotepadImage,
+};
+
+export default notepadImages;
